Functionality added to the "new" settings
The code is not working, just pushed this commit for getting help
