+++ app/src/main/java/com/rmkrings/activities/MainActivity.java
@@ -68,7 +68,7 @@ public class MainActivity extends AppCompatActivity implements ReachabilityChang
                     case R.id.navigation_settings:
                         Intent a = new Intent(MainActivity.this, PreferencesActivity.class);
                         startActivity(a);
-                        return false;
+                        return true;
                 }
                 return false;
             }
