+++ app/src/main/java/com/rmkrings/activities/PreferencesActivity.java
@@ -4,15 +4,17 @@ import android.os.Bundle;
 
 import androidx.appcompat.app.AppCompatActivity;
 import androidx.appcompat.widget.Toolbar;
+import androidx.fragment.app.Fragment;
+import androidx.fragment.app.FragmentTransaction;
 import androidx.viewpager.widget.ViewPager;
 
 import com.google.android.material.tabs.TabLayout;
 import com.rmkrings.data.adapter.ViewPagerAdapter;
+import com.rmkrings.fragments.preferences.IOnBackPressed;
 
 public class PreferencesActivity extends AppCompatActivity {
 
     private ViewPager mViewPager;
-    private Toolbar mToolbar;
     private ViewPagerAdapter mViewPagerAdapter;
     private TabLayout mTabLayout;
 
@@ -20,10 +22,15 @@ public class PreferencesActivity extends AppCompatActivity {
     protected void onCreate(Bundle savedInstanceState) {
         super.onCreate(savedInstanceState);
         setContentView(R.layout.activity_preferences);
-        mToolbar = (Toolbar) findViewById(R.id.toolbar);
-//        setSupportActionBar(mToolbar);
 
         setViewPager();
+    }
+
+    @Override public void onBackPressed() {
+        Fragment fragment = getSupportFragmentManager().findFragmentById(R.id.preferences_general);
+        if(!(fragment instanceof IOnBackPressed)){
+            super.onBackPressed();
+        }
     }
 
     private void setViewPager() {
@@ -34,5 +41,16 @@ public class PreferencesActivity extends AppCompatActivity {
 
         mTabLayout = (TabLayout) findViewById(R.id.tab);
         mTabLayout.setupWithViewPager(mViewPager);
+    }
+
+    private void startFragment(Fragment f, Boolean withBackStack) {
+        FragmentTransaction transaction = getSupportFragmentManager().beginTransaction();
+        transaction.replace(R.id.frameLayout, f);
+
+        if (withBackStack) {
+            transaction.addToBackStack(null);
+        }
+
+        transaction.commit();
     }
 }
