+++ app/src/main/java/com/rmkrings/fragments/DashboardFragment.java
@@ -27,7 +27,7 @@ import android.widget.ImageButton;
 import android.widget.ProgressBar;
 import android.widget.TextView;
 
-import com.rmkrings.activities.SettingsActivity;
+import com.rmkrings.activities.PreferencesActivity;
 import com.rmkrings.data.adapter.DashboardListAdapter;
 import com.rmkrings.data.adapter.MetaDataAdapter;
 import com.rmkrings.data.vertretungsplan.GradeItem;
@@ -173,7 +173,7 @@ public class DashboardFragment extends Fragment implements HttpResponseCallback 
                     .setNegativeButton(R.string.title_settings, new DialogInterface.OnClickListener() {
                         @Override
                         public void onClick(DialogInterface dialog, int which) {
-                            fragmentActivity.startActivity(new Intent(fragmentActivity, SettingsActivity.class));
+                            fragmentActivity.startActivity(new Intent(fragmentActivity, PreferencesActivity.class));
                         }
                     })
                     .show();
