+++ app/src/main/res/layout/fragment_preferences_general.xml
@@ -4,7 +4,8 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".SettingsActivity">
+    tools:context=".PreferencesActivity"
+    android:id="@+id/preferences_general">
 
     <ProgressBar
         android:id="@+id/progressBar"
